Moved the lease factories into the lease test file
This keeps things tidy. If we end up needing leases in other tests we
can always move them back.
